Updating tests to long long int.

+++ tests/math_tests/time_tests/tmpl_math_time_tests.h
@@ -36,7 +36,7 @@ do {                                                                           \
                                                                                \
     const type start = (type)begin;                                            \
     const type end = (type)finish;                                             \
-    const unsigned long int N = (unsigned long int)samples;                    \
+    const unsigned long long int N = (unsigned long long int)samples;          \
     const type dx = (end - start) / (type)N;                                   \
                                                                                \
     x = malloc(sizeof(*x) * N);                                                \
@@ -69,7 +69,7 @@ do {                                                                           \
     printf(#f0 " vs. " #f1 "\n");                                              \
     printf("start:   %.16Le\n", (long double)start);                           \
     printf("end:     %.16Le\n", (long double)end);                             \
-    printf("samples: %lu\n", N);                                               \
+    printf("samples: %llu\n", N);                                              \
     printf("dx:      %.16Le\n", (long double)dx);                              \
                                                                                \
     x[0] = start;                                                              \
